refactor(common) move file interceptors to files module
