+++ packages/common/files/index.ts
@@ -1,5 +1,3 @@
-export * from './file-fields.interceptor';
-export * from './file.interceptor';
-export * from './files.interceptor';
+export * from './interceptors';
 export * from './interfaces';
 export * from './multer.module';
