+++ packages/common/files/interceptors/file-fields.interceptor.ts
@@ -1,19 +1,21 @@
 import * as multer from 'multer';
 import { Observable } from 'rxjs';
-import { Inject, Optional } from '../decorators';
-import { mixin } from '../decorators/core/component.decorator';
-import { ExecutionContext } from '../interfaces';
-import { MulterOptions } from '../interfaces/external/multer-options.interface';
-import { NestInterceptor } from './../interfaces/features/nest-interceptor.interface';
-import { MULTER_MODULE_OPTIONS } from './files.constants';
-import { MulterModuleOptions } from './interfaces';
-import { transformException } from './multer/multer.utils';
+import { Inject, Optional } from '../../decorators';
+import { mixin } from '../../decorators/core/component.decorator';
+import { ExecutionContext } from '../../interfaces';
+import {
+  MulterField,
+  MulterOptions,
+} from '../../interfaces/external/multer-options.interface';
+import { NestInterceptor } from './../../interfaces/features/nest-interceptor.interface';
+import { MULTER_MODULE_OPTIONS } from './../files.constants';
+import { MulterModuleOptions } from './../interfaces';
+import { transformException } from './../multer/multer.utils';
 
 type MulterInstance = any;
 
-export function FilesInterceptor(
-  fieldName: string,
-  maxCount?: number,
+export function FileFieldsInterceptor(
+  uploadFields: MulterField[],
   localOptions?: MulterOptions,
 ) {
   class MixinInterceptor implements NestInterceptor {
@@ -37,7 +39,7 @@ export function FilesInterceptor(
       const ctx = context.switchToHttp();
 
       await new Promise((resolve, reject) =>
-        this.upload.array(fieldName, maxCount)(
+        this.upload.fields(uploadFields)(
           ctx.getRequest(),
           ctx.getResponse(),
           err => {
