+++ packages/common/files/interceptors/file.interceptor.ts
@@ -1,21 +1,18 @@
 import * as multer from 'multer';
 import { Observable } from 'rxjs';
-import { Inject, Optional } from '../decorators';
-import { mixin } from '../decorators/core/component.decorator';
-import { ExecutionContext } from '../interfaces';
-import {
-  MulterField,
-  MulterOptions,
-} from '../interfaces/external/multer-options.interface';
-import { NestInterceptor } from './../interfaces/features/nest-interceptor.interface';
-import { MULTER_MODULE_OPTIONS } from './files.constants';
-import { MulterModuleOptions } from './interfaces';
-import { transformException } from './multer/multer.utils';
+import { Inject, Optional } from '../../decorators';
+import { mixin } from '../../decorators/core/component.decorator';
+import { ExecutionContext } from '../../interfaces';
+import { MulterOptions } from '../../interfaces/external/multer-options.interface';
+import { NestInterceptor } from './../../interfaces/features/nest-interceptor.interface';
+import { MULTER_MODULE_OPTIONS } from './../files.constants';
+import { MulterModuleOptions } from './../interfaces';
+import { transformException } from './../multer/multer.utils';
 
 type MulterInstance = any;
 
-export function FileFieldsInterceptor(
-  uploadFields: MulterField[],
+export function FileInterceptor(
+  fieldName: string,
   localOptions?: MulterOptions,
 ) {
   class MixinInterceptor implements NestInterceptor {
@@ -39,7 +36,7 @@ export function FileFieldsInterceptor(
       const ctx = context.switchToHttp();
 
       await new Promise((resolve, reject) =>
-        this.upload.fields(uploadFields)(
+        this.upload.single(fieldName)(
           ctx.getRequest(),
           ctx.getResponse(),
           err => {
