+++ packages/common/files/interceptors/files.interceptor.ts
@@ -1,18 +1,19 @@
 import * as multer from 'multer';
 import { Observable } from 'rxjs';
-import { Inject, Optional } from '../decorators';
-import { mixin } from '../decorators/core/component.decorator';
-import { ExecutionContext } from '../interfaces';
-import { MulterOptions } from '../interfaces/external/multer-options.interface';
-import { NestInterceptor } from './../interfaces/features/nest-interceptor.interface';
-import { MULTER_MODULE_OPTIONS } from './files.constants';
-import { MulterModuleOptions } from './interfaces';
-import { transformException } from './multer/multer.utils';
+import { Inject, Optional } from '../../decorators';
+import { mixin } from '../../decorators/core/component.decorator';
+import { ExecutionContext } from '../../interfaces';
+import { MulterOptions } from '../../interfaces/external/multer-options.interface';
+import { NestInterceptor } from './../../interfaces/features/nest-interceptor.interface';
+import { MULTER_MODULE_OPTIONS } from './../files.constants';
+import { MulterModuleOptions } from './../interfaces';
+import { transformException } from './../multer/multer.utils';
 
 type MulterInstance = any;
 
-export function FileInterceptor(
+export function FilesInterceptor(
   fieldName: string,
+  maxCount?: number,
   localOptions?: MulterOptions,
 ) {
   class MixinInterceptor implements NestInterceptor {
@@ -36,7 +37,7 @@ export function FileInterceptor(
       const ctx = context.switchToHttp();
 
       await new Promise((resolve, reject) =>
-        this.upload.single(fieldName)(
+        this.upload.array(fieldName, maxCount)(
           ctx.getRequest(),
           ctx.getResponse(),
           err => {
